- There was a reason why I put the metadata in setup.py and not in __init__.py
