+++ setup.py
@@ -1,23 +1,29 @@
+import json
+import os
 from typing import List
 
 import setuptools
 
-from tautulli import (
-    __title__,
-    __version__,
-    __license__,
-    __description__,
-    __author__,
-    __author_email__,
-    __github_username__,
-    __github_repo__,
-    __keywords__,
-    __supported_api_versions__,
-    __supported_python_versions__,
-)
+__version__ = 'VERSIONADDEDBYGITHUB'
+
+__title__ = "tautulli"
+__author__ = 'Nate Harris'
+__author_email__ = 'n8gr8gbln@gmail.com'
+__github_username__ = "nwithan8"
+__github_repo__ = "pytulli"
+__copyright__ = "Copyright © YEARADDEDBYGITHUB - Nate Harris"
+__license__ = 'GNU General Public License v3 (GPLv3)'
+__description__ = "A complete Python client for Tautulli's API"
+__keywords__ = ["Tautulli", "API", "client", "Plex", "PMS", "Plex Media Server", "media", "server", "JSON"]
 
 with open("README.md", "r") as fh:
     long_description = fh.read()
+
+def __supported_python_versions__() -> List[str]:
+    """Return a list of supported Python versions."""
+    with open(os.path.join(os.path.dirname(__file__), "tautulli", "PYTHON_VERSIONS.json")) as f:
+        versions = f.read()
+        return json.loads(versions)
 
 def python_versions() -> List[str]:
     """Return a list of supported Python versions."""
